requirements.txt
fastapi
uvicorn
python-dotenv
python-jose
python-multipart
pydantic